Emailer Function working and secured
