+++ .env
@@ -19,7 +19,9 @@ APP_SECRET=cb0e8700ee36063157863b91c6d9080c
 ###< symfony/framework-bundle ###
 
 ###> symfony/mailer ###
-# MAILER_DSN=smtp://localhost
+MAILER_DSN=gmail+smtp://USERNAME:PASSWORD@default
+DEV_MAILER_RECIPIENT=someone@example.com
+DEV_MAILER_SENDER=someone@example.com
 ###< symfony/mailer ###
 
 ###> doctrine/doctrine-bundle ###
@@ -30,3 +32,8 @@ APP_SECRET=cb0e8700ee36063157863b91c6d9080c
 DATABASE_URL="mysql://db_user:db_password@127.0.0.1:3306/db_name?serverVersion=5.7"
 # DATABASE_URL="postgresql://db_user:db_password@127.0.0.1:5432/db_name?serverVersion=13&charset=utf8"
 ###< doctrine/doctrine-bundle ###
+
+###> symfony/google-mailer ###
+# Gmail SHOULD NOT be used on production, use it in development only.
+# MAILER_DSN=gmail://USERNAME:PASSWORD@default
+###< symfony/google-mailer ###
