+++ composer.json
@@ -26,6 +26,7 @@
         "symfony/flex": "^1.3.1",
         "symfony/form": "5.3.*",
         "symfony/framework-bundle": "5.3.*",
+        "symfony/google-mailer": "5.3.*",
         "symfony/http-client": "5.3.*",
         "symfony/intl": "5.3.*",
         "symfony/mailer": "5.3.*",
