+++ config/packages/gc-communicator.code-workspace
@@ -1,0 +1,8 @@
+{
+	"folders": [
+		{
+			"path": "..\\.."
+		}
+	],
+	"settings": {}
+}
