+++ config/packages/mailer.yaml
@@ -1,3 +1,6 @@
 framework:
     mailer:
         dsn: '%env(MAILER_DSN)%'
+        envelope:
+            sender: '%env(DEV_MAILER_SENDER)%'
+            recipients: ['%env(DEV_MAILER_RECIPIENT)%']
