+++ config/services.yaml
@@ -11,6 +11,10 @@ services:
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
+    Symfony\Component\Mailer\EventListener\EnvelopeListener:
+        arguments:
+            - null
+            - ['%env(DEV_MAILER_RECIPIENT)%']
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
