+++ templates/list_category.html.twig
@@ -8,7 +8,7 @@
   <body>
     {{ include("nav_bar.html.twig") }}
 
-    <div class="w3-panel">
+    <div class="w3-container">
       <h2><b>Categories</b></h2>
       <p>{{ date }}</p>
     </div>
