+++ templates/new_announcement.html.twig
@@ -8,7 +8,7 @@
 
   <body>
     {{ include("nav_bar.html.twig") }}
-    <div class="w3-panel">
+    <div class="w3-container">
       <h2><b>Submit New Announcement</b></h2>
       <p>{{ date }}</p>
     </div>
