+++ templates/new_category.html.twig
@@ -7,7 +7,7 @@
 
   <body>
     {{ include("nav_bar.html.twig") }}
-    <div class="w3-panel">
+    <div class="w3-container">
       <h2><b>Create New Category</b></h2>
       <p>{{ date }}</p>
     </div>
