+++ templates/unauthorized.html.twig
@@ -6,11 +6,9 @@
   </head>
 
   <body>
-    <div class="w3-panel">
+    <div class="w3-container">
       <h2><b>Access Denied</b></h2>
       <p>{{ date }}</p>
-    </div>
-    <div class="w3-panel">
     </div>
   </body>
