+++ templates/user_overview.html.twig
@@ -8,12 +8,12 @@
 
   <body>
     {{ include("nav_bar.html.twig") }}
-    <div class="w3-panel">
+    <div class="w3-container">
       <h2><b>Announcements</b></h2>
       <h3><b>Modify</b></h3>
       <p>{{ date }}</p>
     </div>
-    <div class="w3-panel .dbtheme-purple">
+    <div class="w3-panel">
       <table class="w3-table-all">
         <tr>
           <th>Subject</th>
